+++ src/assets/styles/DefaultLayout.tsx
@@ -1,8 +1,10 @@
+import '../styles/Normalize.css'
+
 interface Props {
-  children: React.ReactNode;
+  children: React.ReactNode
 }
 
 const DefaultLayout: React.FC<Props> = (props: Props) => {
-  return <div>{props.children}</div>;
-};
-export default DefaultLayout;
+  return <div>{props.children}</div>
+}
+export default DefaultLayout
