+++ src/component/common/Header.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useCallback } from 'react'
-import '../../styles/main/Header.scss'
+import '../../styles/common/Header.scss'
 import Modal from './Modal'
-import Login from './Login'
-import SignUp from './SignUp'
+import Login from '../account/Login'
+import SignUp from '../account/SignUp'
 
 const Header: React.FC = () => {
   const [isModalOpen, setModalOpen] = useState<boolean>(false)
