+++ src/styles/account/Login.scss
@@ -1,0 +1,78 @@
+.logInContainer{
+  border: 1px Solid red;
+  width: 100%;
+  height: 100%;
+}
+.loginWrap {
+  border: 1px Solid red;
+  width: 100%;
+  height: 100%;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+}
+.loginTitle {
+  border-radius: 10px;
+  width: 80%;
+  margin: 20px 20px 20px 20px;
+  text-align: center;
+
+  background-color: rgb(184, 184, 184);
+}
+.inputContainer {
+  border: 1px Solid red;
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+
+  div{
+    display: flex;
+    justify-content: space-between;
+    font-size: larger;
+
+    label{
+      width: 150px;
+      height: 50px;
+      text-align: center;
+      line-height: 55px;
+      border-radius: 10px;
+      background-color: rgb(184, 184, 184)
+    }
+
+    input{
+      width: 250px;
+      border-radius: 10px;
+      background-color: rgb(184, 184, 184)
+    }
+  }
+}
+.buttonContainer{
+  border: 1px solid red;
+  width: 100%;
+  flex: 1;
+  
+  button{
+    width: 150px;
+    height: 50px;
+    border-radius: 10px;
+    background-color: rgb(184, 184, 184)
+  }
+}
+
+.socialBottonContainer {
+  border: 1px solid red;
+  width: 80%;
+  display: flex;
+  justify-content: space-around;
+  align-items : center;
+
+  button{
+    width: 150px;
+    height: 50px;
+    border-radius: 10px;
+    background-color: rgb(184, 184, 184)
+  }
+}
