+++ src/styles/common/Modal.scss
@@ -1,6 +1,6 @@
 .modalContainer{
   width: 100%;
-  height: 100%;
+  height: 75%;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -8,8 +8,8 @@
 }
 
 .dialogBox{
-  width: 800px;
-  height: 400px;
+  width: 600px;
+  height: 500px;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -17,7 +17,7 @@
   border-radius: 3px;
   box-shadow: 0 0 30px rgba(30, 30, 30, 0.185);
   box-sizing: border-box;
-  background-color: white;
+  background-color: rgb(162, 162, 162);
   z-index: 10000;
 }
