--- src/styles/main/Login.scss
@@ -1,12 +1,0 @@
-.logInContainer{
-  border: 1px Solid red;
-  width: 100%;
-  height: 100%;
-}
-.loginWrap {
-  border: 1px Solid red;
-  width: 50%;
-  
-  margin: auto;
-  padding: 50px 50px 0px 50px
-}
